requirements.txt
langchain-community
langchain-openai
langchain-classic
langchain-text-splitters
faiss-cpu
sentence-transformers
python-dotenv
